+++ src/umpire/strategy/DynamicPoolList.hpp
@@ -37,6 +37,11 @@ class DynamicPoolList : public AllocationStrategy {
       std::function<bool(const strategy::DynamicPoolList&)>;
 
   static CoalesceHeuristic percent_releasable(int percentage);
+  static CoalesceHeuristic blocks_releasable(std::size_t nblocks);
+
+  static constexpr std::size_t s_default_first_block_size{512 * 1024 * 1024};
+  static constexpr std::size_t s_default_next_block_size{1 * 1024 * 1024};
+  static constexpr std::size_t s_default_alignment{16};
 
   /*!
    * \brief Construct a new DynamicPoolList.
@@ -52,10 +57,9 @@ class DynamicPoolList : public AllocationStrategy {
    */
   DynamicPoolList(
       const std::string& name, int id, Allocator allocator,
-      const std::size_t first_minimum_pool_allocation_size = (512 * 1024 *
-                                                              1024),
-      const std::size_t next_minimum_pool_allocation_size = (1 * 1024 * 1024),
-      const std::size_t alignment = 16,
+      const std::size_t first_minimum_pool_allocation_size = s_default_first_block_size,
+      const std::size_t next_minimum_pool_allocation_size = s_default_next_block_size,
+      const std::size_t alignment = s_default_alignment,
       CoalesceHeuristic should_coalesce = percent_releasable(100)) noexcept;
 
   DynamicPoolList(const DynamicPoolList&) = delete;
@@ -63,6 +67,9 @@ class DynamicPoolList : public AllocationStrategy {
   void* allocate(size_t bytes) override;
   void deallocate(void* ptr, std::size_t size) override;
   void release() override;
+
+  std::size_t getReleasableBlocks() const noexcept;
+  std::size_t getTotalBlocks() const noexcept;
 
   std::size_t getActualSize() const noexcept override;
   std::size_t getCurrentSize() const noexcept override;