+++ tools/replay/ReplayOptions.hpp
@@ -22,18 +22,32 @@ struct ReplayUsePoolValidator : public CLI::Validator {
   }
 };
 
+struct ReplayUseHeuristicValidator : public CLI::Validator {
+  ReplayUseHeuristicValidator() {
+    func_ = [](const std::string &str) {
+      if (str != "Block" && str != "FreePercentage") {
+        return std::string("Invalid heuristic name, must be Block or FreePercentage");
+      }
+      else
+        return std::string();
+    };
+  }
+};
+
 struct ReplayOptions {
   bool time_replay_run{false};    // -t,--time-run
   bool time_replay_parse{false};  // --time-parse
-  bool info_only{false};         // --info
-  bool print_statistics{false};   // -s,--stats
-  bool print_size_stats{false};// --size-stats
+  bool info_only{false};          // --info
+  bool dump_statistics{false};    // -d, --dump
+  bool track_stats{false};        // -s, --stats
   bool skip_operations{false};    // --skip-operations
   bool force_compile{false};      // -r,--recompile
   bool do_not_demangle{false};    // --no-demangle
   bool quiet{false};              // -q,--quiet
   std::string input_file;         // -i,-infile input_file
   std::string pool_to_use;        // -p,--use-pool
+  std::string heuristic_to_use{}; // --use-heuristic
+  int heuristic_parm{2};          // --heuristic-parm
 };
 
 #endif  // REPLAY_ReplayOptions_HPP