+++ tools/replay/replay.cpp
@@ -21,6 +21,7 @@
 #include "ReplayOptions.hpp"
 
 const static ReplayUsePoolValidator ReplayValidPool;
+const static ReplayUseHeuristicValidator ReplayValidHeuristic;
 
 #endif // !defined(_MSC_VER) && !defined(_LIBCPP_VERSION)
 
@@ -41,11 +42,11 @@ int main(int argc, char* argv[])
   app.add_flag("-t,--time-run", options.time_replay_run,
         "Display time information for replay running operations");
 
-  app.add_flag("-s,--stats", options.print_statistics,
-      "Dump ULTRA file containing memory usage stats for each Allocator");
+  app.add_flag("-d,--dump", options.dump_statistics,
+      "Dump ULTRA file containing memory usage trace for each Allocator");
 
-  app.add_flag("--size-stats", options.print_size_stats,
-      "Display pool allocation size statistics");
+  app.add_flag("-s,--stats", options.track_stats,
+      "Track/Display pool allocation size statistics");
 
   app.add_flag("--info-only" , options.info_only,
       "Information about replay file, no actual replay performed");
@@ -60,7 +61,16 @@ int main(int argc, char* argv[])
       "Force recompile replay binary");
 
   app.add_option("-p,--use-pool", options.pool_to_use,
-    "Specify pool to use: List, Map, or Quick")->check(ReplayValidPool);
+    "Specify pool to use: List, Map, or Quick")
+    ->check(ReplayValidPool);
+
+  app.add_option("--use-heuristic", options.heuristic_to_use,
+    "Specify Pool Heuristic to use: Default, Block, or FreePercentage")
+    ->check(ReplayValidHeuristic);
+
+  app.add_option("--heuristic-parm", options.heuristic_parm,
+    "Specify heuristic parameter to use")
+    ->check(CLI::Range(0,100));
 
   CLI11_PARSE(app, argc, argv);